Move utils package functionalities to the step package
