+++ step/platform.go
@@ -1,4 +1,4 @@
-package utils
+package step
 
 import (
 	"fmt"
