+++ step/platform_test.go
@@ -1,4 +1,4 @@
-package utils
+package step
 
 import (
 	"testing"
